requirements.txt
pygame-ce
pygame-gui>=0.6.14
torch>=2.0
matplotlib>=3.5
